Recommend to avoid self-referential text
In the chapter about audience and medium, the IBM Style Guides
recommends to avoid self-referential text like "this chapter" or "this
topic" as much as possible. As this is particularly problematic in
modular documentation where such text can theoretically appear anywhere
in the document hierarchy, I added a vale style to report these as a
suggestion.