Add a warning for words 'sane' and 'sanity'
In the Word Usage section, the IBM Style Guide recommends to avoid the
use of word 'sane' as it might be offensive to neurodiverse people. As
it provides alternatives, I added it to TermsWarning to provide a
warning.

In addition, the IBM Style Guide recommends to avoid using the phrases
'sanity check' and 'sanity test'. There already is a rule for 'sanity
check' but for completenes and to honor the intent behind these
recommendatitons, I added a warning for 'sanity test' and 'sanity' as
well.