Do not suggest to replace "laptop" with "notebook"
The current version of the IBM Style Guide does not prohibit the use of
the word "laptop" in any way and recommends it to refer to mobile
computer that is not a handheld device. On the contrary, the IBM Style
Guide recommends to use the word "notebook" to refer only to an
interface element that resembles a physical notebook with pages and
tabs.