vale-issue-672 - Use double quotes not single quotes in IDs
This PR addresses #672

My regex knowledge leaves a lot to be disired but I made an attempt at adding this rule.

As it relates to AsciiDoc syntax, I added to the AsciiDoc package.

The tests pass when I run `vale .` but this could do with a sanity check.

Any and all feedback welcome :)