Correct AsciiDoc.ImageContainsAltText criteria
If there are multiple entries in the raw array, they are concatenated
into one large regular expression.  In this case, this meant that we
were searching for the regular expression image::.*\[\]image:.*\[\],
which doesn't find images without alt text.

In this commit, we use a single regular expression that will match
both cases: image with one colon, and image with two colons.